Update setup script (probably it could be simplified a lot)

+++ setup.sh
@@ -1,6 +1,10 @@
+#!/bin/bash
 # Set up dotfiles
 
-CURR_DIR=$(dirname "$0")
+CURR_DIR=$(pwd)
+echo "Current dir: $CURR_DIR"
+
+echo "Shell: $SHELL"
 
 # Install packages
 if [[ "$OSTYPE" == "linux-gnu"* ]]; then
@@ -27,7 +31,9 @@ elif [[ "$OSTYPE" == "darwin"* ]]; then
     # Command line tools
     xcode-select --install
     # Install homebrew
-    /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+    if [ -z "brew -v" ]; then
+	/bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
+    fi
 
     INVOKE_PACMAN="brew"
 fi
@@ -54,21 +60,23 @@ fi
 # NOTE: different branches may not contain all the files
 # zshrc:
 ZSH_RC="$CURR_DIR/.zshrc"
-if [ -f "$ZSH_RC" ];then
+if [ -f "$ZSH_RC" ] && [ ! -L "$HOME/.zshrc" ];then
     ln -s "$ZSH_RC" "$HOME/.zshrc"
 fi
 # Source it
 if [ -n "$($SHELL -c "echo $ZSH_VERSION")" ]; then  # zsh is preferred
+    echo "Sourcing ZSHRC"
     source "$HOME/.zshrc"
 fi
 
 # bashrc:
 BASH_RC="$CURR_DIR/.bashrc"
-if [ -f "$BASH_RC" ];then
+if [ -f "$BASH_RC" ] && [ ! -L "$HOME/.bashrc" ];then
     ln -s "$BASH_RC" "$HOME/.bashrc"
 fi
 
-if [ -n "$($SHELL -c 'echo $BASH_VERSION')" ]; then  # bash is preferred
+if [ -n "$($SHELL -c "echo $BASH_VERSION")" ]; then  # bash is preferred
+    echo "Sourcing BASHRC"
     source "$HOME/.bashrc"
 fi
 
@@ -76,13 +84,22 @@ fi
 GHDIR="$HOME/github"
 mkdir -p "$GHREPOS"
 
+if [ -z "$XDG_CONFIG_HOME" ]; then
+    CONFIG_PATH="$HOME/.config"
+else
+    CONFIG_PATH="$XDG_CONFIG_HOME"
+fi
+
 # Nvim config
 if [ -d "$CURR_DIR/nvim" ]; then
-    ln -s "$CURR_DIR/nvim" "$HOME/.config/nvim"
+
+    if [ ! -L "$CONFIG_PATH/nvim" ]; then
+        ln -s "$CURR_DIR/nvim" "$CONFIG_PATH/nvim"
+    fi
 
     echo "Installing Neovim"
     if [[ "$OSTYPE" == "darwin"* ]]; then
-        $INVOKE_PACMAN install nvim
+        $INVOKE_PACMAN install nvim fd
     elif [[ "$OSTYPE" == "linux-gnu"* ]] && [ -f "apt-get -v" ]; then
         # Compile neovim from source
         cd "$GHDIR" || echo "Unable to find GitHub" && exit 1
@@ -94,13 +111,13 @@ if [ -d "$CURR_DIR/nvim" ]; then
         make CMAKE_BUILD_TYPE=RelWithDebInfo
         echo "Building Neovim from source"
         cd build && cpack -G DEB && sudo dpkg -i nvim-linux64.deb
+
+	sudo apt-get install fd-find
     fi
 
     cd "$CURR_DIR"
 
     # Install neovim requirements
-    $INVOKE_PACMAN install fd-find
-
     if [[  "$OSTYPE" == "darwin"* ]]; then
         brew install pngpaste
     else
@@ -113,25 +130,39 @@ if [ -d "$CURR_DIR/nvim" ]; then
     mkdir "$HOME/.virtualenvs"
     cd "$HOME/.virtualenvs"
     python -m venv debugpy
-    debugpy/bin/python -m pip install debugpy
+    source debugpy/bin/activate
+    pip3 install --upgrade pip
+    python3 -m pip install debugpy
+    deactivate
 fi
 
 # Tmux config
 if [ -n "$XDG_CONFIG_HOME" ] && [ -d "$CURR_DIR/tmux" ]; then  # Should be defined in .zshrc
-    ln -s "$CURR_DIR/tmux" "$XDG_CONFIG_HOME/tmux"
+    if [ ! -L "$XDG_CONFIG_HOME/tmux" ]; then
+        ln -s "$CURR_DIR/tmux" "$XDG_CONFIG_HOME/tmux"
+    fi
 elif [ -n "$XDG_CONFIG_HOME" ] && [ -f "$CURR_DIR/.tmux.conf" ]; then
     TMUX_XDG_PATH="$XDG_CONFIG_HOME/tmux"
     mkdir TMUX_XDG_PATH
-    ln -s "$CURR_DIR/.tmux.conf" "$TMUX_XDG_PATH/tmux.conf"
+    if [ ! -L "$TMUX_XDG_PATH/tmux.conf" ]; then
+        ln -s "$CURR_DIR/.tmux.conf" "$TMUX_XDG_PATH/tmux.conf"
+    fi
 elif [ -z "$XDG_CONFIG_HOME" ] && [ -d "$CURR_DIR/tmux" ]; then
-    ln -s "$CURR_DIR/tmux/tmux.conf" "$HOME/.tmux.conf"
+    if [ ! -L "$HOME/.tmux.conf" ]; then
+        ln -s "$CURR_DIR/tmux/tmux.conf" "$HOME/.tmux.conf"
+    fi
 elif [ -z "$XDG_CONFIG_HOME" ] && [ -f "$CURR_DIR/.tmux.conf" ]; then
-    ln -s "$CURR_DIR/.tmux.conf" "$HOME/.tmux.conf"
+    if [ ! -L "$HOME/.tmux.conf" ]; then
+        ln -s "$CURR_DIR/.tmux.conf" "$HOME/.tmux.conf"
+    fi
 else
     echo "No TMUX config found"
 fi
 
 # Vim config
-if [ -f "$CURR_DIR/.vimrc" ]; then
+if [ -f "$CURR_DIR/.vimrc" ] && [ ! -L "$HOME/.vimrc" ]; then
     ln -s "$CURR_DIR/.vimrc" "$HOME/.vimrc"
 fi
+
+echo "Setup complete!"
+exit 0
